// src/app/contract/models/equipment.model.ts
export interface EquipmentType {
  id: string;
  equipmentNumber: string;
  serialNumber: string;
  equipmentDescription: string;
  siteName: string;
  contractNumber: null;
  contractType: null;
  startDate: null;
  endDate: null;
  annualNet: null;
  contractLineItemNumber: null;
  equipmentStatus: string;
  siteZip: string;
  siteState: string;
  partNumber: string;
  systemName: string;
  siteStatus: string;
  endOfSupport: Date;
  modality: string;
  lastUpdatedDate: Date;
  siteCity: string;
  siteStreet: string;
  endOfLife: Date;
  contractLineItemId: null;
  adlsLoadedDate: Date;
  siteName1: string;
  soldToCustomerNumber: string;
  currency: null;
  servicePlan: null;
  shipToCustomerNumber: string;
  technicalIdentificationNumber: string;
  sapModality: null;
  warrantyEndDate: null;
}